requirements.txt
streamlit==1.31.0
Pillow==10.2.0
pandas==2.1.4
openpyxl==3.1.2
google-genai
python-dotenv
